+++ dzgui.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -o pipefail
 
-version=5.6.0-beta.11
+version=5.6.0-beta.16
 
 #CONSTANTS
 aid=221100
@@ -54,7 +54,8 @@ km_helper="$helpers_path/latlon"
 sums_path="$helpers_path/sums.md5"
 func_helper="$helpers_path/funcs"
 
-#URLS
+#REMOTE
+remote_host=gh
 author="aclist"
 repo="dztui"
 url_prefix="https://raw.githubusercontent.com/$author/$repo"
@@ -63,6 +64,7 @@ testing_url="$url_prefix/testing"
 releases_url="https://github.com/$author/$repo/releases/download/browser"
 km_helper_url="$releases_url/latlon"
 geo_file_url="$releases_url/ips.csv.gz"
+
 
 set_im_module(){
     #TODO: drop pending SteamOS changes
@@ -312,13 +314,14 @@ check_unmerged(){
 check_version(){
     local version_url=$(format_version_url)
     local upstream=$(curl -Ls "$version_url" | awk -F= '/^version=/ {print $2}')
+    local res=$(get_response_code "$version_url")
+    [[ $res -ne 200 ]] && raise_error_and_quit "Remote resource unavailable: '$version_url'"
     logger INFO "Local branch: '$branch', local version: $version"
     if [[ $branch == "stable" ]]; then
         version_url="$stable_url/dzgui.sh"
     elif [[ $branch == "testing" ]]; then
         version_url="$testing_url/dzgui.sh"
     fi
-    local upstream=$(curl -Ls "$version_url" | awk -F= '/^version=/ {print $2}')
     [[ ! -f "$freedesktop_path/$app_name.desktop" ]] && freedesktop_dirs
     if [[ $version == $upstream ]]; then
         logger INFO "Local version is same as upstream"
@@ -371,9 +374,10 @@ prompt_dl(){
 }
 dl_changelog(){
     local mdbranch
+    local md
     [[ $branch == "stable" ]] && mdbranch="dzgui"
     [[ $branch == "testing" ]] && mdbranch="testing"
-    local md="https://raw.githubusercontent.com/$author/dztui/${mdbranch}/CHANGELOG.md"
+    local md="$url_prefix/${mdbranch}/$file"
     curl -Ls "$md" > "$state_path/CHANGELOG.md"
 }
 test_display_mode(){
@@ -515,6 +519,7 @@ get_hash(){
     md5sum "$1" | awk '{print $1}'
 }
 fetch_a2s(){
+    # this file is currently monolithic
     [[ -d $helpers_path/a2s ]] && { logger INFO "A2S helper is current"; return 0; }
     local sha=c7590ffa9a6d0c6912e17ceeab15b832a1090640
     local author="yepoleb"
@@ -522,6 +527,8 @@ fetch_a2s(){
     local url="https://github.com/$author/$repo/tarball/$sha"
     local prefix="${author^}-$repo-${sha:0:7}"
     local file="$prefix.tar.gz"
+    local res=$(get_response_code "$url")
+    [[ $res -ne 200 ]] && raise_error_and_quit "Remote resource unavailable: '$file'"
     curl -Ls "$url" > "$helpers_path/$file"
     tar xf "$helpers_path/$file" -C "$helpers_path" "$prefix/a2s" --strip=1
     rm "$helpers_path/$file"
@@ -535,9 +542,11 @@ fetch_dzq(){
         return 0
     fi
     local sha=3088bbfb147b77bc7b6a9425581b439889ff3f7f
-    local author="aclist"
+    local author="yepoleb"
     local repo="dayzquery"
     local url="https://raw.githubusercontent.com/$author/$repo/$sha/dayzquery.py"
+    local res=$(get_response_code "$url")
+    [[ $res -ne 200 ]] && raise_error_and_quit "Remote resource unavailable: 'dayzquery.py'"
     curl -Ls "$url" > "$file"
     logger INFO "Updated DZQ to sha '$sha'"
 }
@@ -572,7 +581,7 @@ fetch_helpers_by_sum(){
         ["ui.py"]="be3da1e542d14105f4358dd38901e25a"
         ["query_v2.py"]="55d339ba02512ac69de288eb3be41067"
         ["vdf2json.py"]="2f49f6f5d3af919bebaab2e9c220f397"
-        ["funcs"]="62f6b3fb2dcb56a78b7642c0f0aa7abe"
+        ["funcs"]="10c7d9cb9fbb792626ec9e7a4a788ba5"
         ["lan"]="c62e84ddd1457b71a85ad21da662b9af"
     )
     local author="aclist"
@@ -596,11 +605,15 @@ fetch_helpers_by_sum(){
         file="$i"
         sum="${sums[$i]}"
         full_path="$helpers_path/$file"
-        url="https://raw.githubusercontent.com/$author/$repo/$realbranch/helpers/$file"
+
+        url="${url_prefix}/$realbranch/helpers/$file"
+
         if [[ -f "$full_path" ]] && [[ $(get_hash "$full_path") == $sum ]]; then
             logger INFO "$file is current"
         else
             logger WARN "File '$full_path' checksum != '$sum'"
+            local res=$(get_response_code "$url")
+            [[ $res -ne 200 ]] && raise_error_and_quit "Remote resource unavailable: '$url'"
             curl -Ls "$url" > "$full_path"
             if [[ ! $? -eq 0 ]]; then
                 raise_error_and_quit "Failed to fetch the file '$file'. Possible timeout?"
@@ -621,11 +634,15 @@ fetch_geo_file(){
     local km_sum="b038fdb8f655798207bd28de3a004706"
     local gzip="$helpers_path/ips.csv.gz"
     if [[ ! -f $geo_file  ]] || [[ $(get_hash $geo_file) != $geo_sum ]]; then
+        local res=$(get_response_code "$geo_file_url")
+        [[ $res -ne 200 ]] && raise_error_and_quit "Remote resource unavailable: '$geo_file_url'"
         curl -Ls "$geo_file_url" > "$gzip"
         #force overwrite
         gunzip -f "$gzip"
     fi
     if [[ ! -f $km_helper ]] || [[ $(get_hash $km_helper) != $km_sum ]]; then
+        local res=$(get_response_code "$km_helper_url")
+        [[ $res -ne 200 ]] && raise_error_and_quit "Remote resource unavailable: '$km_helper_url'"
         curl -Ls "$km_helper_url" > "$km_helper"
         chmod +x "$km_helper"
     fi
@@ -845,14 +862,42 @@ is_steam_running(){
         return 0
     fi
 }
+get_response_code(){
+    local url="$1"
+    curl -Ls -I -o /dev/null -w "%{http_code}" "$url"
+}
 test_connection(){
-    ping -c1 -4 github.com 1>/dev/null 2>&1
-    if [[ ! $? -eq 0 ]]; then
-        raise_error_and_quit "No connection could be established to the remote server (github.com)."
+    source "$config_file"
+    declare -A hr
+    local res1
+    local res2
+    local str="No connection could be established to the remote server"
+    hr=(
+        ["steampowered.com"]="https://api.steampowered.com/IGameServersService/GetServerList/v1/?key=$steam_api"
+        ["github.com"]="https://github.com/$author"
+        ["codeberg.org"]="https://codeberg.org/$author"
+    )
+    # steam API is mandatory, except on initial setup
+    if [[ -n $steam_api ]]; then
+        res=$(get_response_code "${hr["steampowered.com"]}")
+        [[ $res -ne 200 ]] && raise_error_and_quit "$str ("steampowered.com")"
     fi
-    ping -c1 -4 api.steampowered.com 1>/dev/null 2>&1
-    if [[ ! $? -eq 0 ]]; then
-        raise_error_and_quit "No connection could be established to the remote server (steampowered.com)."
+
+    res=$(get_response_code "${hr["github.com"]}")
+    if [[ $res -ne 200 ]]; then
+        logger WARN "Remote host '${hr["github.com"]}' unreachable', trying fallback"
+        remote_host=cb
+        res=$(get_response_code "${hr["codeberg.org"]}")
+        [[ $res -ne 200 ]] && raise_error_and_quit "$str (${hr["codeberg.org"]})"
+    fi
+    logger INFO "Set remote host to '${hr["codeberg.org"]}'"
+    if [[ $remote_host == "cb" ]]; then
+        url_prefix="https://codeberg.org/$author/$repo/raw/branch"
+        releases_url="https://codeberg.org/$author/$repo/releases/download/browser"
+        stable_url="$url_prefix/dzgui"
+        testing_url="$url_prefix/testing"
+        km_helper_url="$releases_url/latlon"
+        geo_file_url="$releases_url/ips.csv.gz"
     fi
 }
 legacy_cols(){